fix: cambio de path de archivos
