+++ pom.xml
@@ -80,7 +80,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.3</version>
                 <configuration>
-                    <mainClass>poo.coders.main.vista.Main</mainClass>
+                    <mainClass>poo.coders.main.Main</mainClass>
                 </configuration>
             </plugin>
         </plugins>
