+++ src/main/resources/groupChoice.fxml
